Various API changes including:
1. Move isSatisifed into RuleEvaluator to minimise the API surface.
2. Add a new global settings for skipping integrity component for rule
providers.
3. Incorporate changes suggested by API review in http://b/147835536.

Note that this change is forked from ag/9974531.

Test: atest frameworks/base/core/tests/coretests/src/android/content/integrity/
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/integrity
Bug: 145290357, 147835536

+++ services/tests/servicestests/src/com/android/server/integrity/utils/TestUtils.java
@@ -18,8 +18,8 @@ package com.android.server.integrity.utils;
 
 public class TestUtils {
 
-    public static String getBits(int component, int numOfBits) {
-        return String.format("%" + numOfBits + "s", Integer.toBinaryString(component))
+    public static String getBits(long component, int numOfBits) {
+        return String.format("%" + numOfBits + "s", Long.toBinaryString(component))
                 .replace(' ', '0');
     }
